+++ packages/core/field-base/docs/0-intro.js
@@ -1,25 +1,21 @@
 // @flow
 import React from 'react';
-import { md, Example, Props } from '@atlaskit/docs';
+import { code, md, Example, Props } from '@atlaskit/docs';
 
 export default md`
-  ### Usage
 
-  This component contains all the common behaviour and styles for fields
+This component contains all the common behaviour and styles for fields.
 
-  ~~~js
-  import FieldBase, {
-    FieldBaseStateless,
-    Label
-  } from '@atlaskit/field-base';
-  ~~~
+FieldBase provides an Atlassian Design Guidelines compatible implementation for:
+* Labels: spacing, margins, accessibility.
+* Fields: sizing, borders, colors, wrapping behaviour, hover/focus states.
+* Validation: styles (built in validation coming soon!)
 
-  FieldBase provides an Atlassian Design Guidelines compatible implementation for:
-  * Labels: spacing, margins, accessibility.
-  * Fields: sizing, borders, colors, wrapping behaviour, hover/focus states.
-  * Validation: styles (built in validation coming soon!)
+## Usage
 
-  FieldBase components *will* work by themselves, but are really meant to be extended into a full field component.
+${code`import FieldBase, {FieldBaseStateless,Label} from '@atlaskit/field-base';`}
+
+FieldBase components *will* work by themselves, but are really meant to be extended into a full field component.
 
   ${(
     <Example