+++ packages/core/navigation-next/src/index.js
@@ -31,9 +31,6 @@ export {
 export {
   default as PeekToggleItem,
 } from './components/presentational/PeekToggleItem';
-export {
-  default as ScrollableSectionInner,
-} from './components/presentational/ScrollableSectionInner';
 export { default as Section } from './components/presentational/Section';
 export {
   default as SectionHeading,
@@ -58,6 +55,9 @@ export { default as GoToItem } from './components/connected/GoToItem';
 export {
   default as LayoutManagerWithViewController,
 } from './components/connected/LayoutManagerWithViewController';
+export {
+  default as AsyncLayoutManagerWithViewController,
+} from './components/connected/AsyncLayoutManagerWithViewController';
 export { default as SortableGroup } from './components/connected/SortableGroup';
 export { default as SortableItem } from './components/connected/SortableItem';
 export {
@@ -79,7 +79,10 @@ export {
 } from './view-controller';
 
 /** Renderer */
-export { default as ViewRenderer } from './renderer';
+export {
+  default as ViewRenderer,
+  TypedItemsRenderer as TypedViewRenderer,
+} from './renderer';
 
 /** Theme */
 export { dark, light, settings, modeGenerator, ThemeProvider } from './theme';
@@ -105,3 +108,28 @@ export type {
   WithGlobalThemeProps,
   GlobalThemeWrappedComp,
 } from './theme/types';
+
+export type {
+  NavigationRendererItemType,
+  LeafItemType,
+  BranchItemType,
+  BackItemType,
+  ContainerHeaderType,
+  DebugType,
+  GoToItemType,
+  GroupHeadingType,
+  ItemType,
+  SortableItemType,
+  SectionHeadingType,
+  SeparatorType,
+  SwitcherType,
+  WordmarkType,
+  GroupType,
+  HeaderSectionType,
+  MenuSectionType,
+  SectionType,
+  SortableContextType,
+  SortableGroupType,
+  InlineComponentType,
+  TypeShape as NavigationItemTypeShape,
+} from './renderer/types';