+++ packages/editor/adf-utils/src/__tests__/unit/validator/validate.ts
@@ -21,7 +21,7 @@ const readFilesSync = (path: string) =>
   );
 
 describe('validate', () => {
-  ['full', 'stage-0'].forEach(schemaType => {
+  ['stage-0'].forEach(schemaType => {
     const valid = readFilesSync(`${BASE_DIR}/${schemaType}/valid`);
     valid.forEach(file => {
       // Don't test Application Card