+++ packages/editor/adf-utils/src/validator/specs/heading.ts
@@ -8,5 +8,6 @@ export default {
       optional: true,
     },
     attrs: { props: { level: { type: 'number', minimum: 1, maximum: 6 } } },
+    marks: { type: 'array', items: ['alignment'], optional: true },
   },
 };